Move TruePointcut to the correct namespace
